this is a valid change for debugging failed tests
svn:revision:48922
svn:branch:trunk
svn:account:tbrowder2

+++ regress/mged.sh
@@ -156,20 +156,22 @@ else
 fi
 
 # clean up
-# remove test databases
-tgms="mged.g"
-for t in $tgms ; do
-  if test -f $t ; then
-    rm $t
-  fi
-done
-# remove test files
-tfils="t.solids t.regions"
-for t in $tfils ; do
-  if test -f $t ; then
-    rm $t
-  fi
-done
+# remove test databases (but only if tests succeed)
+if test $FAILED -ne 0 ; then
+    tgms="mged.g"
+    for t in $tgms ; do
+      if test -f $t ; then
+        rm $t
+      fi
+    done
+    # remove test files
+    tfils="t.solids t.regions"
+    for t in $tfils ; do
+      if test -f $t ; then
+        rm $t
+      fi
+    done
+fi
